+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 0.9 *(not released)*
 * Fixed issue for windows environments, not containing "PATH" but "Path" _(tspaeth)_
 * Allow 64 bit windows to use the x64 node.exe _(ekaufman)_
 * Renamed "ext" property on SetupTask so that it's not causing any conflicts
+* Added detection for FreeBSD as a Linux variant
 * ...
 
 Version 0.8 *(2014-11-19)*